Paginate by 50 records
I noticed that pagination was only going one record at a time even though it
was displaying 50 records on each page. For example, if the home page showed
the first 50 records, clicking next showed 50 records starting from the
second record. I forgot to multiply offset by recordsPerPage which is a
value that I named instead of hardcoding 50 in multiple places